Получение объекта для работы с YouTube API выделено в отдельный класс Getyoutube (getyoutube.py), классы Channel и Video являются по отношению к нему дочерними (подклассами класса Getyoutube). Внесены изменения в указанные классы. Проверена работоспособность кода.
