+++ src/data/moscowpython.json
@@ -3,7 +3,7 @@
   "title": "MoscowPython",
   "description": "Видеозаписи со встреч питонистов и джангистов в Москве и не только. :)\nПрисоединяйтесь: https://www.facebook.com/groups/MoscowDjango! :)",
   "url": "https://www.youtube.com/channel/UC-OVMPlMA3-YCIeg4z5z23A",
-  "subscriberCount": "26600",
-  "videoCount": "709",
-  "viewCount": "2434644"
+  "subscriberCount": "26700",
+  "videoCount": "714",
+  "viewCount": "2444785"
 }
